app/requirements.txt
boto3==1.12.39
botocore==1.15.39
click==7.1.1
docutils==0.15.2
Flask==1.1.2
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
numpy==1.18.2
packaging==20.3
pandas==0.24.2
protobuf==3.11.3
protobuf3-to-dict==0.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
s3transfer==0.3.3
sagemaker==1.55.3
scipy==1.4.1
six==1.14.0
smdebug-rulesconfig==0.1.2
urllib3==1.25.8
Werkzeug==1.0.1
zipp==3.1.0
